requirements.txt
# Real Crypto Twitter Scraper Requirements
selenium==4.18.1
webdriver-manager>=4.0.1
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.4
lxml>=5.3.0
pyarrow>=14.0.1  # for parquet support
python-dotenv==1.0.1
supabase==2.3.5
